Working dynamic implementation.
The program now iteriates through all of the images found in the directory and outputs the results of the matches between all of the images. The iteration number is also appended to the end of the images, as well as the matcher type. The results are categorized in folders, so ORB goes to ORB folder, etc.